Restructured project
Put things in compositor/, utils/, and plugins/ respectively. They
should be fairly self explanatory.

Also finally put headers in a top level include that mirrors the
projects setup.
