+++ include/compositor/cursor.h
