+++ way-cooler/compositor/meson.build
@@ -1,0 +1,13 @@
+way_cooler_sources += files(
+	'cursor.c',
+	'input.c',
+	'keyboard.c',
+	'layer_shell.c',
+	'output.c',
+	'pointer.c',
+	'seat.c',
+	'server.c',
+	'view.c',
+	'xdg.c',
+	'xwayland.c',
+)
