+++ way-cooler/plugins/meson.build
@@ -1,0 +1,3 @@
+way_cooler_sources += files(
+	'keybindings.c',
+)
