+++ way-cooler/utils/meson.build
@@ -1,0 +1,4 @@
+way_cooler_sources += files(
+	'xkb_hash_set.c',
+)
+
