+++ way-cooler/utils/xkb_hash_set.c
@@ -1,4 +1,4 @@
-#include "xkb_hash_set.h"
+#include "utils/xkb_hash_set.h"
 
 #include <stdlib.h>
 #include <string.h>
